requirements.txt
asgiref==3.4.1
crispy-forms-materialize==0.2
Django==3.2.7
django-allauth==0.45.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-image-optimizer==1.0.1
django-localflavor==3.1
django-multiselectfield==0.1.12
oauthlib==3.1.1
Pillow==8.3.2
PyJWT==2.1.0
python-resize-image==1.1.19
python-stdnum==1.16
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
stripe==2.60.0
tinify==1.5.2
